add API function to remove existing references in user-objects when deleting a recipe

+++ pages/recipe/[id]/index.js
@@ -61,6 +61,9 @@ export default function DetailPage({
   };
 
   async function handleDelete() {
+    if (confirm("Dieses Rezept löschen?") !== true) {
+      return;
+    }
     const response = await fetch(`/api/recipes/${id}`, {
       method: "DELETE",
       headers: {
